uma par de coisa

+++ Perfumaria/Models/Perfumes.cs
@@ -5,7 +5,8 @@
         public int id {  get; set; }
         public string nome { get; set; }
         public string descricao { get; set; }
-        public float valor { get; set; }
+        //public Categoria idCategoria { get; set; }
+        public decimal valor { get; set; }
         public int quantidade { get; set; }
     }
 }
